transform/requirements.txt
pandas
s3fs
python-dotenv
sqlalchemy
psycopg2-binary